service/requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Login==0.2.6
Flask-SQLAlchemy==2.1
boto3==1.3
psycopg2
uwsgi